Added new class to use aioharmony instead of pyharmony
The main advantage is that aioharmony uses the websockets API, which is
also used by the Harmony iPhone app.  In theory, it'll be better
supported than the XMPP API which is considered "dangerous" and is
unsupported by Logitech.  Not to mention it's disabled by default for
versions of the firmware since December 2018.